requirements.txt
discord.py[voice]
asyncio
aiosqlite
python-dotenv